Move Zones package out of the Viewer package
