+++ engine/src/main/java/org/terasology/world/generation/WorldBuilder.java
@@ -23,7 +23,7 @@ import com.google.common.collect.Sets;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.terasology.world.generator.plugin.WorldGeneratorPluginLibrary;
-import org.terasology.world.viewer.zones.Zone;
+import org.terasology.world.zones.Zone;
 
 import java.util.ArrayList;
 import java.util.Arrays;
